reorganize PropertyPane and PropertyGroup
