--- src/components/atoms/PropertyPane/index.stories.tsx
@@ -1,10 +1,0 @@
-import React from "react";
-import { Meta } from "@storybook/react";
-import PropertyPane from ".";
-
-export default {
-  title: "atoms/PropertyPane",
-  component: PropertyPane,
-} as Meta;
-
-export const Default = () => <PropertyPane />;
